Move runtime information helpers from execution_sycl_defs.h to parallel_backend_sycl_utils.h
