+++ include/oneapi/dpl/pstl/hetero/dpcpp/parallel_backend_sycl_utils.h
@@ -29,6 +29,7 @@
 #include "../../iterator_impl.h"
 
 #include "sycl_defs.h"
+#include "execution_sycl_defs.h"
 #include "sycl_iterator.h"
 #include "../../utils.h"
 
@@ -43,6 +44,109 @@ namespace oneapi
 {
 namespace dpl
 {
+namespace __internal
+{
+
+//-----------------------------------------------------------------------------
+// Device run-time information helpers
+//-----------------------------------------------------------------------------
+
+#if _ONEDPL_DEBUG_SYCL
+template <typename _ExecutionPolicy>
+::std::string
+__device_info(_ExecutionPolicy&& __policy)
+{
+    return __policy.queue().get_device().template get_info<sycl::info::device::name>();
+}
+#endif
+
+template <typename _ExecutionPolicy>
+::std::size_t
+__max_work_group_size(_ExecutionPolicy&& __policy)
+{
+    return __policy.queue().get_device().template get_info<sycl::info::device::max_work_group_size>();
+}
+
+template <typename _ExecutionPolicy, typename _Size>
+_Size
+__adjust_to_local_mem_size(_ExecutionPolicy&& __policy, _Size __local_mem_per_wi, _Size __wg_size)
+{
+    auto __local_mem_size = __policy.queue().get_device().template get_info<sycl::info::device::local_mem_size>();
+    return ::std::min(__local_mem_size / __local_mem_per_wi, __wg_size);
+}
+
+#if _USE_SUB_GROUPS
+template <typename _ExecutionPolicy>
+::std::size_t
+__max_sub_group_size(_ExecutionPolicy&& __policy)
+{
+    auto __supported_sg_sizes = __policy.queue().get_device().template get_info<sycl::info::device::sub_group_sizes>();
+
+    //The result of get_info<sycl::info::device::sub_group_sizes>() can be empty - the function returns 0;
+    return __supported_sg_sizes.empty() ? 0 : __supported_sg_sizes.back();
+}
+#endif
+
+template <typename _ExecutionPolicy>
+::std::uint32_t
+__max_compute_units(_ExecutionPolicy&& __policy)
+{
+    return __policy.queue().get_device().template get_info<sycl::info::device::max_compute_units>();
+}
+
+//-----------------------------------------------------------------------------
+// Kernel run-time information helpers
+//-----------------------------------------------------------------------------
+
+// 20201214 value corresponds to Intel(R) oneAPI C++ Compiler Classic 2021.1.2 Patch release
+#define _USE_KERNEL_DEVICE_SPECIFIC_API (__SYCL_COMPILER_VERSION > 20201214) || (_ONEDPL_LIBSYCL_VERSION >= 50700)
+
+template <typename _ExecutionPolicy>
+::std::size_t
+__kernel_work_group_size(_ExecutionPolicy&& __policy, const sycl::kernel& __kernel)
+{
+    const sycl::device& __device = __policy.queue().get_device();
+    const ::std::size_t __max_wg_size =
+#if _USE_KERNEL_DEVICE_SPECIFIC_API
+        __kernel.template get_info<sycl::info::kernel_device_specific::work_group_size>(__device);
+#else
+        __kernel.template get_work_group_info<sycl::info::kernel_work_group::work_group_size>(__device);
+#endif
+    // The variable below is needed to achieve better performance on CPU devices.
+    // Experimentally it was found that the most common divisor is 4 with all patterns.
+    // TODO: choose the divisor according to specific pattern.
+    ::std::size_t __cpu_divisor = 1;
+    if (__device.is_cpu() && __max_wg_size >= 4)
+        __cpu_divisor = 4;
+
+    return __max_wg_size / __cpu_divisor;
+}
+
+template <typename _ExecutionPolicy>
+::std::uint32_t
+__kernel_sub_group_size(_ExecutionPolicy&& __policy, const sycl::kernel& __kernel)
+{
+    const sycl::device& __device = __policy.queue().get_device();
+    [[maybe_unused]] const ::std::size_t __wg_size =
+        __kernel_work_group_size(::std::forward<_ExecutionPolicy>(__policy), __kernel);
+    const ::std::uint32_t __sg_size =
+#if _USE_KERNEL_DEVICE_SPECIFIC_API
+        __kernel.template get_info<sycl::info::kernel_device_specific::max_sub_group_size>(
+            __device
+#    if _ONEDPL_LIBSYCL_VERSION < 60000
+            ,
+            sycl::range<3> { __wg_size, 1, 1 }
+#    endif
+        );
+#else
+        __kernel.template get_sub_group_info<sycl::info::kernel_sub_group::max_sub_group_size>(
+            __device, sycl::range<3>{__wg_size, 1, 1});
+#endif
+    return __sg_size;
+}
+
+} // namespace __internal
+
 namespace __par_backend_hetero
 {
